backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
groq==0.5.0
supabase==2.4.2
python-dotenv==1.0.1
pydantic==2.6.4
boto3==1.34.69
mangum==0.17.0
httpx==0.27.0
Pillow==10.3.0
imageio[ffmpeg]==2.34.1
numpy==1.26.4
PyPDF2==3.0.1
python-docx==1.1.2
python-multipart==0.0.9
